backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
python-dotenv==1.2.2
requests==2.32.5
aiohttp==3.13.3
python-docx==1.2.0
yandexcloud==0.384.0
email-validator==2.3.0
boto3==1.42.83
langfuse==4.0.6
pdfplumber==0.11.9
loguru==0.7.3
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
pytest-mock==1.1.0
python-magic==0.4.27
faster-whisper==1.2.1
ollama==0.6.1
nvidia-cublas-cu12
nvidia-cudnn-cu12
nvidia-cuda-runtime-cu12
torch

